Avoid testing missing backends if possible
git-svn-id: svn://svn.code.sf.net/p/hfst/code/trunk/hfst3@904 941e2c2b-deac-454f-805a-451daa25f33c

+++ test/tools/conjunct-functionality.sh
@@ -1,10 +1,12 @@
 #!/bin/sh
 for i in "" .sfst .ofst .foma; do
-if ! ../../tools/src/hfst-conjunct cat_or_dog.hfst$i cat.hfst$i > test.hfst ; then
-    exit 1
-fi
-if ! ../../tools/src/hfst-compare test.hfst cat.hfst$i > /dev/null 2>&1 ; then
-    exit 1
-fi
-rm test.hfst;
+    if test -f cat_or_dog.hfst$i -a -f cat.hfst$i ; then
+        if ! ../../tools/src/hfst-conjunct cat_or_dog.hfst$i cat.hfst$i > test.hfst ; then
+            exit 1
+        fi
+        if ! ../../tools/src/hfst-compare test.hfst cat.hfst$i > /dev/null 2>&1 ; then
+            exit 1
+        fi
+        rm test.hfst;
+    fi
 done
