+++ test/tools/fst2strings-functionality.sh
@@ -1,10 +1,12 @@
 #!/bin/sh
 for i in "" .sfst .ofst .foma; do
-if ! ../../tools/src/hfst-fst2strings < cat.hfst$i > test.strings ; then
-    exit 1
-fi
-if ! diff test.strings cat.strings > /dev/null 2>&1 ; then
-    exit 1
-fi
-rm test.strings;
+    if test -f cat.hfst$i ; then
+        if ! ../../tools/src/hfst-fst2strings < cat.hfst$i > test.strings ; then
+            exit 1
+        fi
+        if ! diff test.strings cat.strings > /dev/null 2>&1 ; then
+            exit 1
+        fi
+        rm test.strings;
+    fi
 done
